Working Imputer Feature-Gen, but slow
The Imputer Feature-Generator is working, but really slow in this
commit. I think that's because currently, for every generated feature,
the imputer model is called. And these calls I suspect are really slow.

So this needs to be improved! I'll try that going forward. This is a
safe heaven though to start working from :)