Refactor pdep feature generator
The pdep vicinity feature generator now returns only one feature. The
feature's probability is the column's gpdep score times the conditional
probability of the value used to correct the error.

It might be necessary to normalize these values to make them compatible
with the other feature generators -- the domain feature generator for
example returns only two probabilities: 1.0, or 0.